Fix init.el loaded in bkg when slow tag skipped
Starting the service would force loading the (slow) bootstrap script
as part of its init, but as background task (systemd service) instead
of as explicitly-requested foreground task.

This means even if we skipped the @slow tag above, the config would be
loaded (albeit in background).

Since the systemd service is already enabled, it'll be loaded on next
boot anyway, no need to activate it now: removed.