docs: Update README for sergei
README was out of whack now that I've transformed this repo into a
multi-machine setup.

Presence of Sergei (the name of the server machine) in the repo means
every vagrant statement needs to describe what machine it's intended
for.

Might be better to end up splitting sergei into its own Vagrantfile
and folder...