emacs: Fix exec-path not loaded on startup
Was only loading when GUI is present, despite being inside the non-X11
package config, loaded before any GUI is present.

Remove the unnecessary check (which is described on package homepage
but not really justified) to fix the loading. If I find a good reason
for keeping the check (such as extra envvars set in X11 that need
loading) I can always add a re-invokation of the initialize function
inside X11 conf file.