server: Enable install-lock on gitea config
Fixes the major issue with the gitea config, which was trying to write
modifications to the config file at boot time for some reason.

No. I like immutable system and its config. Enable that flag. Doesn't
seem to have specific impact, maybe it was the secret tokens I set to
hardcoded values?