Change default branch to main
It's shorter to type, and it's even more accurate than "master"
regarding what the branch IS: the MAIN branch, not the maestro
(as in music mixing) branch.

Despite initially disliking the change as frivolous (and it is/was),
I've lived with it for a couple months at work and I kind of like it,
really is 1/3rd shorter to type.