i3: Fix keyboard FR layout missing+capslock rebind
Issue with layout seems to be some other startup commands are overriding
the setxkbmap call we're doing. Such commands unidentified as of yet.

Worked around via explicit sleeps to the command, ensuring we're
artificially the last one to run it.

Should investigate some way to detect what command is the one doing it,
and remove the sleep, which pauses such thing for 10s